ajout du mode CUDA
