+++ Makefile
@@ -1,7 +1,7 @@
 all:
-	meson setup bin/
 	meson compile -C bin/
 run:
-	meson setup bin/
 	meson compile -C bin/
 	bin/main
+configure:
+	meson setup bin/ --reconfigure
