+++ meson.build
@@ -1,2 +1,3 @@
-project('nombres-premiers', 'c')
-executable('main', 'Code/main.c')
+project('nombres-premiers', 'c', version : '1.0.0')
+# thread = dependency('treads')
+executable('premiers', 'Code/CPU/main.c', install : true)
